more data loading error handling improvements
